// src/seed/logic/reports/shared-types.ts
export enum ELogicResStatusCode {
  /**codigo inicial de grupo informativo*/
  INFO = 10,
  /**codigo de recepcion de  informacion
   * para el usuario*/
  INFO_USER = 11,
  /**Peticion procesandose */
  PROCESSING = 12,
  /** codigo inicial de grupo cuando
   * una peticion o ejecucion de accion
   * fue exitosa*/
  SUCCESS = 20,
  /** Codigo que indica que la informacion
   * registrada por el usuario es valida*/
  VALID_DATA = 21,
  /**codigo inicial de grupo de advertencia recibida */
  WARNING = 25,
  /**codigo que indica que la informacion
   * de usuario presenta advertencias */
  WARNING_DATA = 26,
  /**redireccionamiento*/
  REDIRECT = 30,
  /**codigo inicial de grupo de accion
   * invalida recibida*/
  BAD = 40,
  /**codigo que indica no tiene autorizacion */
  INVALID_USER = 41,
  /**codigo que indica que el usuario ingresó
   * informacion invalida que no se puede
   * almacenar o consultar */
  INVALID_DATA = 44,
  /**codigo inicial de grupo de error
   * en algun modulo que impide la
   * cualquier ejecucion*/
  ERROR = 50,
}
// ...
export enum EKeyActionGroupForRes {
  //==== contexto Profundos ============================================
  /**indica grupo de subacciones ejecutadas*/
  actions = "#ACTIONS#",
  /**indica grupo de subacciones ejecutadas de las
   *  propiedades correspondientes a un objeto anonimo
   */
  props = "#PROPS#",
  /**indica grupo de subacciones ejecutadas de las
   *  propiedades correspondientes a un array de
   * elementos (cualquiera que sea)
   */
  items = "#ITEMS#",
  /**indica grupo de subacciones ejecutadas de las
   *  campos correspondientes a un modelo
   * ❕Exclusivo para contexto estructurado❕
   */
  fields = "#FIELDS#",
  /**indica grupo de subacciones ejecutadas de las
   *  modelos embebidos correspondientes a un modelo
   * ❕Exclusivo para contexto estructurado❕
   */
  embModel = "#EMB_MODELS#",
  //==== contexto Modulos ============================================
  /**indica grupo de sub acciones ejecutadas en el modulo
   * *controller*, del contexto *primitive* */
  ctrlPrimitive = "#CTRL_PRIMITIVE#",
  /**indica grupo de sub acciones ejecutadas en el modulo
   * *controller*, del contexto *structure*, en el contexto
   * profundo *field* */
  ctrlField = "#CTRL_FIELD#",
  /**indica grupo de sub acciones ejecutadas en el modulo
   * *controller*, del contexto *structure*, en el contexto
   * profundo *model* */
  ctrlModel = "#CTRL_MODEL#",
  /**indica grupo de sub acciones ejecutadas en el modulo
   * *mutate*, del contexto *primitive* */
  mutatePrimitive = "#MUTATE_PRIMITIVE#",
  /**indica grupo de sub acciones ejecutadas en el modulo
   * *mutate*, del contexto *structure* */
  mutateStructure = "#MUTATE_STRUCTURE#",
  /**indica grupo de sub acciones ejecutadas en el modulo
   * *mutate*, del contexto *structure*, en el contexto
   * profundo *field* */
  mutateField = "#MUTATE_FIELD#",
  /**indica grupo de sub acciones ejecutadas en el modulo
   * *mutate*, del contexto *structure*, en el contexto
   * profundo *model* */
  mutateModel = "#MUTATE_MODEL#",
  /**indica grupo de sub acciones ejecutadas en el modulo
   * *validation*, del contexto *primitive* */
  valPrimitive = "#VAL_PRIMITIVE#",
  /**indica grupo de sub acciones ejecutadas en el modulo
   * *mutate*, del contexto *structure* */
  valStructure = "#VAL_STRUCTURE#",
  /**indica grupo de sub acciones ejecutadas en el modulo
   * *validation*, del contexto *structure*, en el contexto
   * profundo *field* */
  valField = "#VAL_FIELD#",
  /**indica grupo de sub acciones ejecutadas en el modulo
   * *validation*, del contexto *structure*, en el contexto
   * profundo *model* */
  valModel = "#VAL_MODEL#",
  /**indica grupo de sub acciones ejecutadas en el modulo
   * *hook*, del contexto *primitive* */
  hookPrimitive = "#HOOK_PRIMITIVE#",
  /**indica grupo de sub acciones ejecutadas en el modulo
   * *hook*, del contexto *structure* */
  hookStructure = "#HOOK_STRUCTURE#",
  /**indica grupo de sub acciones ejecutadas en el modulo
   * *provider*, del contexto *primitive* */
  providerPrimitive = "#PROVIDER_PRIMITIVE#",
  /**indica grupo de sub acciones ejecutadas en el modulo
   * *provider*, del contexto *structure* */
  providerStructure = "#PROVIDER_STRUCTURE#",
  /**indica grupo de sub acciones realizadas desde drivers */
  driver = "#DRIVER#",
}
/** */
export interface IResponse {
  /**datos recibidos */
  data: any; //❗Puede ser cualquier dato❗
  /**clave identificadora del modulo */
  keyRepModule: TKeyActionModule;
  /**clave identificadora del contexto lógico que se usó para la petición */
  keyRepLogicContext: TKeyLogicContext;
  /**clave identificadora del contexto de modulo interno */
  keyRepModuleContext: unknown;
  /**clave identificadora del recurso*/
  keyRepSrc: string;
  /**clave identificadora del recurso(modelo o campo), sin ruta */
  keyLogic?: string;
  /**clave identificadora del tipo de request */
  keyTypeRequest: TKeyRequestType;
  /**clave identificadora con contexto de modificacion (solo en peticiones de tipo "modify") */
  keyModifyTypeRequest?: TKeyRequestModifyType;
  /**clave identificadora de la accion que ejecutó el middleware
   *
   * ❕Existen acciones especiales que agrupan subAcciones, no
   * pertenecen a ningun diccionario dfe acciones y estan nombradas
   * en mayuscula y comienzan y terminan con el caracter `"#"` ❕
   */
  keyAction: string | EKeyActionGroupForRes;
  /**clave identificadora de l apeticion de request */
  keyActionRequest: string;
  /**el dato inicial con el que se inició la peticion en el controller*/
  firstCtrlData?: any;
  /**estado despues de la ejecucion del middleware */
  status: ELogicResStatusCode;
  /**mensaje auxiliar (Lógico) */
  msn: string;
  /**contiene un reporte producido por
   * libreria o mudulo externo a la logica */
  extResponse?: object;
  /**tolerancia maxima de respuesta
   * con la que se evaluo este middleware*/
  tolerance: ELogicResStatusCode;
  /**reportes embebidos si los hay */
  responses: IResponse[];
}
// ...
export interface IDriverResponse
  extends Pick<IResponse, "data" | "status" | "msn"> {
  /**detalles adicionales de la respuesta (normalmente entregados por la api externa) */
  details?: any;
  /**detalles específicos del error */
  error?: any;
}